moved emergency stop to nav panel
